Fix narrowing warning in V3Expand::newSelBitBit
VL_BITBIT_E on a signed toSInt() value returns a signed int, which
narrowed implicitly into AstConst's uint32_t brace-init parameter.
Cast explicitly, matching the pattern already used in newWordIndex().

+++ src/V3Expand.cpp
@@ -294,7 +294,8 @@ class ExpandVisitor final : public VNVisitor {
         // holding a negative lsb still yields the correct bit-within-word.
         FileLine* const fl = lsbp->fileline();
         if (VN_IS(lsbp, Const)) {
-            return new AstConst{fl, VL_BITBIT_E(VN_AS(lsbp, Const)->toSInt())};
+            return new AstConst{
+                fl, static_cast<uint32_t>(VL_BITBIT_E(VN_AS(lsbp, Const)->toSInt()))};
         } else {
             return new AstAnd{fl, new AstConst{fl, VL_EDATASIZE - 1}, lsbp->cloneTreePure(true)};
         }
